rewrite Trajectory
Trajectory used to be an abstract base class, instantiated into
subclasses like `Trajectory_1N2d` etc. Refactored to just have a
Trajectory class that deals with everything, since the subclassing
mechanism was overkill / impractical for the usecase.

Updated the rest of the library & tests to be compatible with this
change. Total code coverage is 98%, with 48 of the 92 missing statements
occuring in tracklib.analysis.plots, which will be refactored in the
future anyways.